move MapByVulnID to its own file
Leaving store.go focused on interface definition.
